requirements.txt
pynput
mss
Pillow